Изменил тип в калькуляторе на decimal.

+++ FinanceManagementApp/FinanceManagementAppCore/Entities/Calculator.cs
@@ -4,10 +4,10 @@ namespace Domain.Entities
 {
     public static class Calculator
     {
-        public static double Calc(string expression)
+        public static decimal Calc(string expression)
         {
             Expression e = new Expression(expression);
-            return Convert.ToDouble(Math.Round((double)e.Evaluate(), 2));
+            return Convert.ToDecimal(Math.Round((decimal)e.Evaluate(), 2));
         }
 
         public static double CalcTotalWithPercent(double sum, int percent)
